requirements.txt
openai>=1.11.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
zenpy>=2.0.24
